+++ migrations/0006_variable_secrets.sql
@@ -1,0 +1,4 @@
+-- version: 18
+-- description: Add secret flag to variables
+
+ALTER TABLE variables ADD COLUMN secret INTEGER NOT NULL DEFAULT 0;